requirements.txt
absl-py==2.1.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.14.0
bidict==0.23.1
boto3==1.34.49
botocore==1.34.49
CacheControl==0.12.14
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.1
confluent-kafka==2.3.0
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dask==2022.10.2
debugpy==1.8.0
decorator==5.1.1
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fbmessenger==6.0.0
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
jax==0.4.24
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.3
jsonschema==4.17.3
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.15.0
kiwisolver==1.4.5
labelImg==1.8.6
libclang==16.0.6
locket==1.0.0
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
mattermostwrapper==2.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==5.2.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pamqp==3.2.1
pandas==2.2.0
parso==0.8.3
partd==1.4.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
portalocker==2.8.2
prompt-toolkit==3.0.28
protobuf==4.23.4
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
Pygments==2.17.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
questionary==1.10.0
randomname==0.1.5
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.27.0
SQLAlchemy==1.4.51
stack-data==0.6.3
structlog==23.3.0
structlog-sentry==2.0.3
sympy==1.12
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.13.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.3.0
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
twilio==8.2.2
typing-utils==0.1.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
